Rewrote a lot of code related to calling and performing moves to make the code more readable and flexible
Short leash is also a condition that summons companions at the start of combat.

special sauce and wild form use performMoveAction and wild form also uses companionTakeTurnAction at the end of it.

StaminaPower now has an instant trigger, and everything should be polished to have as few visual bugs as possible.

added lots of comments to make things more readable.